+++ tests/Fixtures/Resources/PostLocalesResource/Pages/CreatePostLocales.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rankbeam\Seo\Filament\Tests\Fixtures\Resources\PostLocalesResource\Pages;
+
+use Filament\Resources\Pages\CreateRecord;
+use Rankbeam\Seo\Filament\Tests\Fixtures\Resources\PostLocalesResource;
+
+class CreatePostLocales extends CreateRecord
+{
+    protected static string $resource = PostLocalesResource::class;
+}